fix(agents): replace chat UI pill with capability icons
