+++ apps/emdash-desktop/src/core/features/settings/browser/agents-page/AgentRow.tsx
@@ -1,7 +1,5 @@
-import { Tooltip } from '@emdash/ui/react/primitives';
-import { MessageSquare, SquareTerminal } from 'lucide-react';
-import type { ReactNode } from 'react';
 import { getAgentUpdateActionState } from '@core/features/agents/api/browser/components/agent-selector/agent-install';
+import { AgentCapabilityIcons } from '@core/features/agents/contributions/browser/agent-capability-icons';
 import { AgentIcon } from '@core/features/agents/contributions/browser/agent-icon';
 import { UpdateAvailableBadge } from '@core/features/settings/browser/agents-page/agent-status-badge';
 import { agentSupportsAcp, type AgentPayload } from '@core/primitives/agents/api';
@@ -30,30 +28,9 @@ export const AgentRow = ({ agent }: { agent: AgentPayload }) => {
         <span className="text-sm text-foreground">{agent.name}</span>
         <div className="flex items-center gap-2">
           {updateState.render && <UpdateAvailableBadge />}
-          {/* Every agent runs in the terminal; ACP-capable ones also offer the chat UI. */}
-          {supportsChatUi && (
-            <CapabilityIcon label="Supports Chat UI">
-              <MessageSquare className="size-3.5 text-foreground-passive" aria-hidden />
-            </CapabilityIcon>
-          )}
-          <CapabilityIcon label="Supports TUI">
-            <SquareTerminal className="size-3.5 text-foreground-passive" aria-hidden />
-          </CapabilityIcon>
+          <AgentCapabilityIcons supportsChatUi={supportsChatUi} />
         </div>
       </div>
     </div>
   );
 };
-
-function CapabilityIcon({ label, children }: { label: string; children: ReactNode }) {
-  return (
-    <Tooltip.Provider delay={150}>
-      <Tooltip.Root>
-        <Tooltip.Trigger render={<span className="inline-flex" aria-label={label} />}>
-          {children}
-        </Tooltip.Trigger>
-        <Tooltip.Content>{label}</Tooltip.Content>
-      </Tooltip.Root>
-    </Tooltip.Provider>
-  );
-}
